requirements.txt
jinja2>=3.0.0
pyyaml>=6.0
python-dotenv>=0.19.0
requests>=2.26.0